ProofOfConcept.sh v1
This is the first official release of my proof of concept for FOAR705. This code will download annotation information on specific website from the hypothes.is API, reformat that information so that it is human readable, it will then analyse the annotation information and provide a list of the most common words used within the user, uri, text, exact, and tag sections. The end result will be two files. One named <filename>.txt which contains the new formatting and analysis, and one named <filename>-original.txt which contains the original annotation information downloaded from the API. To use you will need to type bash ProofOfConcept.sh <URL> <filename>