Update README.md
This addition to the README includes the purpose, usage and results of the proof of concept. The section on purpose explains why hypothes.is is useful and its limitations, which leads me to the purpose of my tool. The usage section explains how to use the tool through terminal and the command bash ProofOfConcept.sh <URL> <filename>. The result explains the two files that will be the end product of the code.