Initial Draft of PICO script
This speech has three sections. The first explains the importance of hypothes.is. The second explains the faults with hypothes.is that my proof of concept will solve. The third explains what my proof of concept is. At the moment the timing of this script is too fast and the script may be changed. However, the timing issues may be due to the lack of slides and the fact that I am currently reading the script rather than presenting from memory.